requirements.txt
behave
python-vagrant
paramico
